refactor: sort files to different folders
