+++ src/components/header/Header.tsx
@@ -6,9 +6,9 @@ import { useLocation } from 'react-router';
 
 import Logo from '../logo/Logo.tsx';
 import Navigation from '../navigation/Navigation.tsx';
-import DropDownPanel from './DropDownPanel.tsx';
-import IconsStack from './Icons-stack.tsx';
-import LoginRegisterButton from './Login-register-button.tsx';
+import DropDownPanel from './components/drop-down-panel/DropDownPanel.tsx';
+import IconsStack from './components/icon-stack/Icons-stack.tsx';
+import LoginRegisterButton from './components/login-register-button/Login-register-button.tsx';
 
 const StyledHeader = styled('header')(({ theme }) => ({
   backgroundColor: theme.palette.background.default,
