+++ src/components/header/components/magnifier/Magnifier.test.tsx
