+++ src/components/header/components/magnifier/Magnifier.tsx
