+++ src/pages/shop/components/searcher/Searcher.tsx
@@ -4,7 +4,7 @@ import TextField from '@mui/material/TextField';
 import { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router';
 
-import Magnifier from '@/components/header/magnifier/Magnifier';
+import Magnifier from '@/components/header/Magnifier/Magnifier';
 
 const Searcher = () => {
   const [query, setQuery] = useState('');
